# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Event Booking & Staff Management API
  description: API for managing events, staff scheduling, participant registration, and communications.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
security:
  - bearerAuth: []
...
paths:
...
  /tags:
    get:
      tags: [Tags]
      summary: List all tags
      responses:
        '200':
          description: List of tags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tag'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    post:
      tags: [Tags]
      summary: Create a new tag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagRequest'
      responses:
        '201':
          description: Tag created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /tags/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    put:
      tags: [Tags]
      summary: Update a tag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagRequest'
      responses:
        '200':
          description: Tag updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '404':
          description: Tag not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    delete:
      tags: [Tags]
      summary: Delete a tag
      responses:
        '204':
          description: Tag deleted
        '404':
          description: Tag not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden

  # ── Locations (Admin & Configuration) ──
...
  /staff:
    get:
      tags: [Staff Management]
      summary: List all staff members (Admin only)
      responses:
        '200':
          description: List of staff
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Staff'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    post:
      tags: [Staff Management]
      summary: Create a new staff account (Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StaffCreateRequest'
      responses:
        '201':
          description: Staff created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Staff'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /staff/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [Staff Management]
      summary: Get staff details
      responses:
        '200':
          description: Staff details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Staff'
        '404':
          description: Staff not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    put:
      tags: [Staff Management]
      summary: Update staff details (Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StaffUpdateRequest'
      responses:
        '200':
          description: Staff updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Staff'
        '404':
          description: Staff not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    delete:
      tags: [Staff Management]
      summary: Delete a staff member (Admin only)
      responses:
        '204':
          description: Staff deleted
        '404':
          description: Staff not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden

  # ── Staff Self-Service ──
  /staff/{id}/availability:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    put:
      tags: [Staff Self-Service]
      summary: Update staff weekly availability windows
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AvailabilityRequest'
      responses:
        '200':
          description: Availability updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvailabilityRequest'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /staff/{id}/max-hours:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    put:
      tags: [Staff Self-Service]
      summary: Update staff maximum working hours per week
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaxHoursRequest'
      responses:
        '200':
          description: Max hours updated
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /staff/{id}/tags:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    put:
      tags: [Staff Self-Service]
      summary: Update staff qualifications/tags
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StaffTagsRequest'
      responses:
        '200':
          description: Tags updated
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /staff/{id}/schedule:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [Staff Self-Service]
      summary: Get upcoming assigned events for the staff member
      responses:
        '200':
          description: Staff schedule
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventInstance'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /staff/{id}/holidays:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [Staff Self-Service]
      summary: List staff holiday dates
      responses:
        '200':
          description: List of holidays
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Holiday'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    post:
      tags: [Staff Self-Service]
      summary: Add a staff holiday date
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HolidayRequest'
      responses:
        '201':
          description: Holiday added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Holiday'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    delete:
      tags: [Staff Self-Service]
      summary: Remove a staff holiday date
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HolidayRequest'
      responses:
        '204':
          description: Holiday removed
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden

  # ── Event Instances ──
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        details:
          type: array
          items:
            type: string
...
    Tag:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          maxLength: 100
    TagRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 100

    # ── Locations ──
...
    Staff:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          maxLength: 150
        email:
          type: string
          format: email
          maxLength: 200
        maxHoursPerWeek:
          type: integer
          description: Maximum working hours per week (maps to users.max_hours_per_week)
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        availability:
          type: array
          items:
            $ref: '#/components/schemas/AvailabilityWindow'
    StaffCreateRequest:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          maxLength: 150
        email:
          type: string
          format: email
          maxLength: 200
        password:
          type: string
        maxHoursPerWeek:
          type: integer
          default: 40
    StaffUpdateRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 150
        email:
          type: string
          format: email
          maxLength: 200
    AvailabilityWindow:
      type: object
      properties:
        dayOfWeek:
          type: string
          enum: [MONDAY, TUESDAY, WEDNESDAY, THURSDAY, FRIDAY, SATURDAY, SUNDAY]
        startTime:
          type: string
          format: time
          example: "09:00"
        endTime:
          type: string
          format: time
          example: "17:00"
    AvailabilityRequest:
      type: object
      required: [windows]
      properties:
        windows:
          type: array
          items:
            $ref: '#/components/schemas/AvailabilityWindow'
    MaxHoursRequest:
      type: object
      required: [maxHoursPerWeek]
      properties:
        maxHoursPerWeek:
          type: integer
          default: 40
    StaffTagsRequest:
      type: object
      required: [tagIds]
      properties:
        tagIds:
          type: array
          items:
            type: integer
            format: int64
    Holiday:
      type: object
      properties:
        id:
          type: integer
          format: int64
        date:
          type: string
          format: date
    HolidayRequest:
      type: object
      required: [date]
      properties:
        date:
          type: string
          format: date

    # ── Event Instances ──
    EventStatus:
      type: string
      description: >
        Expanded to match the relational model (event_instances.status) while keeping API needs.
      enum: [DRAFT, PUBLISHED, STAFFED, COMPLETED, CANCELLED]
    EventInstance:
      type: object
      properties:
        id:
          type: integer
          format: int64
        eventTypeId:
          type: integer
          format: int64
        eventTypeName:
          type: string
        locationId:
          type: integer
          format: int64
        locationName:
          type: string
        date:
          type: string
          format: date
        startTime:
          type: string
          format: time
        endTime:
          type: string
          format: time
        status:
          $ref: '#/components/schemas/EventStatus'
        assignedStaff:
          type: array
          items:
            $ref: '#/components/schemas/Staff'
        maxParticipants:
          type: integer
          description: Capacity override for this instance (maps to event_instances.capacity_override)
        registeredParticipants:
          type: integer
          description: Count of registered participants (derived from registrations table in implementation)
...